♻️ Refactor : refactor getCategory function
